2wab8YOH Minimal updates to cover page
Removed instruction to download the VSP because this is now in the 
tutorial.

Remove information about using the testing service instead of the Hub 
because this is now in the tutorial itself.

Update bullet points that describe what the tutorial will cover.

Call it a 'step-by-step guide' as opposed to 'tutorial'.