RML48JfS Housekeeping tasks
Add next review due date.
Going with quarterly review for now to see how this goes.

Moved links from the bottom of the page to the links rference file.

Removing 'Next steps' because the content it refers to is work in progress.
Will put 'Next steps' back in once the content is published.

Change folder names to have the slug for this page be `/get-started/handle-failure-scenarios`