Re-write Development Steps as a top-level page
The information is the same as in the currently published version.

Re-orgnanised content to focus around a tasklist for each stage.

Removed recurring links asking users to familiarise themseslves with PKI.

Kept content as high-level overview, removing details such as creating and deleting test users.

Removed information about Compliance Tool, which is not visible to the user in the way it used to be.