Reorganise content into component-based sections
User research showed that having all sets of instructions on the same 
page can be confusing.

This is why we decided to break up the content into separate pages.

Having one page per set of instructions leads to 6 pages with long 
titles. The tiles are long because they need to start with a verb, 
mention the component, and specify which type of key is being updated.

The solution at the moment is to have three pages, one for each 
component. Then each page will have the instructions for updating the 
encryption and signing keys.

This commit creates the new page structure and moves the existing 
content in the correct places. Follow-up commits will iterate the 
content.