Remove link to self-service tool
Removing link because link leads to sign-in page.

User research shows this link is not helpful and even confusing for 
users who are at a later stage of the process.

At the moment we can't link specifically to where they are in the 
process so leaving the text unlinked.

User research has also shown users open the docs in a separate tab so 
providing a route back to the self service tool is not a concern.