Draft content for generating self-signed certs
Offering OpenSSL example of generating the .pk8 and .crt files for the 
component config.

The cert naming convention was there to help humans working on the 
Verify side make sense of things.

The Self-Service Admin tool will not requite human involvement so there 
is no more need for a cert naming convention. Keeping a bulleted list 
with what could be useful to have in the cert common name (CN) and why.

This also makes life easier for services who probably have their own 
internal naming conventions.