Add explanation of fields in the MSA config
We saw in user research that users were confused about what information 
they should put in the `name` field when configuring MSA certificates.

The tech team advised it's a name for that certificate and it get 
published to the MSA's metadata together with the certificate.

Add a table after the config file explaining what all the different 
fields are and where the values are used.