Rephrased introduction and turned into list
Following recommendations from 2i, introduction paragraph changed so that failure scenarios are explained in a bullet point list. 

Slight changes to wording of introduction to make it clearer that users need to request to be sent responses from the Verify Hub.

Removed references to 'configuration changes' to improve readability and reduce potential confusion.