Improve content about security incident recommendations
Change bulleted list to numbered list because the order is not crucial - 
both things need to happen as soon as possible.

Say that the Verify team can help reduce the impact of the incident.

Fix phrasing to say you need to rotate keys and certs if your keys are 
compromised.

Refer to the thing as 'GOV.UK Verify Manage certificates service'