Have separate sections for uploading wrong encription and signing keys
The previous content didn't take into account the different actions 
users should take and didn't specifically address sitations when this 
might break the connection.

Replacing works only for encryption so I kept this, and added info about 
what happens if the wrong encryption cert is uploaded.

For signing they need to remove the wrong one and upload a new one. 
Simply uploading the new one won't work and the docs need to make this 
clear.