Write troubleshooting section
We need a more comprehensive and user centred section about how to 
troubleshoot a broken connection as a result of key rotation.

Starting from the experience of devs on support, we're developing a 
step-by step troubleshooting procedure first addressing the most common 
points of failure and then diving into key-cert mismatches, which are 
actaully quite rare.