Rephrase intro to Step 3
Rephrasing intro to Step 3 to avoid using a numbered list. This is a 
good idea because:

There are 2 other numbered lists in this section and each of them tells 
the user to go to the Manage certificates service.

It's a bit of an overkill to have a 3 step list for something that can 
be achieved by linking to a page where users should easily be able to 
make a choice.