Swap step order for cert uploading and config changing - MSA signing
The swap does not break anytihng, the cert uploading and config changing
can happen in any order.

The old order is:
2. upload new cert
3. change MSA config

The new order is:
2. change MSA config
3. upload new cert

We are making the change because then the steps would be similar to the
MSA encryption flow and users could do both at the same time. This means
they could get away with 4 deployments rather than 2. This came out of
the private beta testing sesions.

This commit also:
* adds indenting to ERB conditionals
* moves the paragraph about (V)SP picking up new MSA metadata to the
beginning of the delete key step, where the user is more likely to read
it.