Rewrite and rename the partial for getting certs
The partial for getting certs is now `get-first-certs` because the 
content is only used when connecting to Integration and Production. 
`get-first certs` includes content from `get-self-signed-certificates`, 
which is also used in key rotation instructions.

The content has been re-written to match the new setup where services 
use private keys and self-signed certificates to secure messages 
echanged with the GOV.UK Verify Hub.

The guidance for storing private keys was kept because that remains 
relevant for the setup using self-signed certificates.

Updated the name of the partial in the files where it gets called.