Change config/tech-docs.yml to match master
When we started prototyping, we made changes to config/tech-docs.yml 
because we were running a paralel build for testing. For example, that 
build was not allowed to have the GOV.UK logo, it had the PROTOTYPE 
label, and we also didn't call it "Verify documentation" to avoid 
confusion.

This commit applies the configuration from the master branch to avoid 
merge conflicts.